minor bug fix + submodule updates
fixed a very minor bug where writing cpu/vpu debug strings wasn't clearing the stream's width format setting.
fixed by setting width to zero before generating the format restore sentry.

updated the asm and ios-frstor modules.

fixed warning where we were implicitly casting u64 to u8.

+++ src/AsmArgs.cpp
@@ -2212,7 +2212,7 @@ bool AssembleArgs::__TryProcessMOVxX_settings_byte(bool sign, u64 dest, u64 dest
 	default: res = {AssembleError::UsageError, "line " + tostr(line) + ": Specified size combination is not supported"}; return false;
 	}
 
-	return TryAppendByte((dest << 4) | mode);
+	return TryAppendByte((u8)((dest << 4) | mode));
 }
 bool AssembleArgs::TryProcessMOVxX(OPCode opcode, bool sign)
 {
